modules: add clear_cmd functionality to several modules

+++ AstrakoBot/modules/zombie.py
@@ -1,26 +1,26 @@
 from asyncio import sleep
-from telegram import Bot, Update, ParseMode
-from telegram.ext import Updater, CommandHandler
-from telegram.ext import CallbackContext, run_async
 from telethon import events
 from AstrakoBot import dispatcher
 from AstrakoBot import telethn as AstrakoBotTelethonClient
+from AstrakoBot.modules.sql.clear_cmd_sql import get_clearcmd
 from AstrakoBot.modules.helper_funcs.telethn.chatstatus import user_is_admin
+from AstrakoBot.modules.helper_funcs.misc import delete
 
 
 @AstrakoBotTelethonClient.on(events.NewMessage(pattern=f"^[!/]zombies ?(.*)"))
 async def zombies(event):
     chat = await event.get_chat()
+    chat_id = event.chat_id
     admin = chat.admin_rights
     creator = chat.creator
 
     if not await user_is_admin(
         user_id = event.sender_id, message = event
     ):
-        msg = f"Only Admins are allowed to use this command"
+        delmsg = "Only Admins are allowed to use this command"
 
     elif not admin and not creator:
-        msg = f"I am not an admin here!"
+        delmsg = "I am not an admin here!"
 
     else:
 
@@ -28,34 +28,37 @@ async def zombies(event):
         arg = event.pattern_match.group(1).lower()
 
         if not arg:
-                msg = f"**Searching for zombies...**\n"
+                msg = "**Searching for zombies...**\n"
+                msg = await event.reply(msg)
                 async for user in event.client.iter_participants(event.chat):
                     if user.deleted:
                         count += 1
 
                 if count == 0:
-                    msg += f"No deleted accounts found. Group is clean"
+                    delmsg = await msg.edit("No deleted accounts found. Group is clean")
                 else:
-                    msg += f"Found **{count}** zombies in this group\n"
-                    msg += f"Clean them by using - `/zombies clean`"
+                    delmsg = await msg.edit(f"Found **{count}** zombies in this group\nClean them by using - `/zombies clean`")
         
         elif arg == "clean":
-            msg = f"**Cleaning zombies...**\n"
+            msg = "**Cleaning zombies...**\n"
+            msg = await event.reply(msg)
             async for user in event.client.iter_participants(event.chat):
                 if user.deleted:
                     count += 1
                     await event.client.kick_participant(chat, user)
 
             if count == 0:
-                msg += f"No zombies account found. Group is clean"
+                delmsg = await msg.edit("No deleted accounts found. Group is clean")
             else:
-                msg += f"Cleaned `{count}` zombies"
+                delmsg = await msg.edit(f"Cleaned `{count}` zombies")
       
         else:
-            msg = "Wrong parameter. You can use only `/zombies clean`"
+            delmsg = await event.reply("Wrong parameter. You can use only `/zombies clean`")
 
 
-    delmsg = await event.reply(msg)
-    await sleep(60)
-    await delmsg.delete()
+    cleartime = get_clearcmd(chat_id, "zombies")
+
+    if cleartime:
+        await sleep(cleartime.time)
+        await delmsg.delete()
